backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
psycopg[binary]
python-dotenv
passlib[bcrypt]
bcrypt<4
python-jose[cryptography]
cryptography
paramiko
itsdangerous
pydantic[email]
pydantic-settings
websockets
python-multipart
pytest
pytest-asyncio
httpx
